feat(openamt): rename packages
