+++ api/hostmanagement/intel/openamt/authorization.go
