+++ api/hostmanagement/intel/openamt/configDomain.go
